feat: [web-components] test separate package
