+++ packages/puik-web-components/components/button-group.ts
@@ -1,0 +1,6 @@
+import { defineCustomElement } from 'vue';
+import { PuikButtonGroup } from '@prestashopcorp/puik-components';
+
+const PuikButtonGroupCe = defineCustomElement(PuikButtonGroup);
+
+export default PuikButtonGroupCe;
