+++ packages/puik-web-components/components/button.ts
@@ -1,0 +1,6 @@
+import { defineCustomElement } from 'vue';
+import { PuikButton } from '@prestashopcorp/puik-components';
+
+const PuikButtonCe = defineCustomElement(PuikButton);
+
+export default PuikButtonCe;
