+++ packages/puik-web-components/components/icon.ts
@@ -1,0 +1,6 @@
+import { defineCustomElement } from 'vue';
+import { PuikIcon } from '@prestashopcorp/puik-components';
+
+const PuikIconCe = defineCustomElement(PuikIcon);
+
+export default PuikIconCe;
