+++ packages/puik-web-components/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@headlessui/vue": "^1.7.11",
     "@popperjs/core": "^2.11.6",
-    "@prestashopcorp/puik-theme": "workspace:*",
+    "@prestashopcorp/puik-components": "workspace:*",
+    "@prestashopcorp/puik-tailwind-preset": "workspace:*",
     "@vueuse/core": "^9.13.0"
   },
   "devDependencies": {
