+++ packages/puik-web-components/vite.config.ts
@@ -23,7 +23,8 @@ export default defineConfig({
       external: [
         ...Object.keys(pkg.dependencies),
         ...Object.keys(pkg.peerDependencies),
-        /^@prestashopcorp\/puik-theme\/.*/
+        /^@prestashopcorp\/puik-theme\/.*/,
+        /^@prestashopcorp\/puik-components\/.*/
       ],
       input: excludeFiles(
         await glob('./**/*.{vue,ts}', {
