+++ packages/puik/index.ts
@@ -1,6 +1,5 @@
 import installer from './defaults';
 export * from '@prestashopcorp/puik-components';
-export * from '@prestashopcorp/puik-web-components';
 export { PuikResolver } from '@prestashopcorp/puik-resolver';
 export { puikTailwindPreset } from '@prestashopcorp/puik-tailwind-preset';
