+++ packages/puik/package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@prestashopcorp/puik-components": "workspace:*",
-    "@prestashopcorp/puik-web-components": "workspace:*",
     "@prestashopcorp/puik-resolver": "workspace:*",
     "@prestashopcorp/puik-tailwind-preset": "workspace:*",
     "@prestashopcorp/puik-theme": "workspace:*"
